algorithms: tiny memory space savings by changing bitness

+++ src/algorithms/grade.go
@@ -1,6 +1,6 @@
 package algorithms
 
-func CalculateGrade(score uint64) string {
+func CalculateGrade(score uint32) string {
 	switch {
 	case score >= 9900000:
 		return "S"
